hwwf to 4 v1
