+++ simple-chat/package.json
@@ -2,7 +2,8 @@
   "private": true,
   "scripts": {
     "dev": "cross-env NODE_ENV=development webpack --mode development",
-    "build": "cross-env NODE_ENV=production webpack --mode production"
+    "build": "cross-env NODE_ENV=production webpack --mode production",
+    "start": "webpack-dev-server --mode development"
   },
   "devDependencies": {
     "@babel/core": "^7.25.8",
@@ -15,9 +16,11 @@
     "css-loader": "^7.1.2",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.6.0",
+    "imask": "^7.6.1",
     "mini-css-extract-plugin": "^2.9.1",
     "style-loader": "^4.0.0",
     "webpack": "^5.95.0",
-    "webpack-cli": "^5.1.4"
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.1.0"
   }
 }
