+++ simple-chat/src/index.html
@@ -19,6 +19,20 @@
                 <div class="chatList">
                     
                 </div>
+                <div class="configSide">
+                    <div class="navigation_Panel">
+                        <div class="add_Navigation">
+                            <span class="material-symbols-outlined">
+                                add
+                            </span>
+                        </div>
+                        <div class="user_Navigation">
+                            <span class="material-symbols-outlined">
+                                account_circle
+                            </span>
+                        </div>
+                    </div>
+                </div>
             </div>
             <div class="rightSide">
                 <div class="header">
@@ -39,7 +53,9 @@
                 </div>
                 <div class="chatBox">
                     <div class="chatWrapper" id="chatWrapper" name="1">
-                        <div class="wrapper_Scrollbar"></div>
+                        <div class="conversationSection">
+                            <div class="wrapper_Scrollbar"></div>
+                        </div>
                     </div>
                 </div>
 
@@ -49,8 +65,10 @@
                         <input type="file" class="attachMedia">
                     </div>
                     <input type="text" class="input" placeholder="Сообщение">
-                    <div class="sendArrow material-symbols-outlined">
-                        send
+                    <div class="sendArrow">
+                        <span class="material-symbols-outlined">
+                            send
+                        </span>
                     </div>
                 </div>
             </div>
