+++ simple-chat/src/index.js
@@ -1,2 +1,7 @@
-import './styles.css';
+import './main.css'
+import './styles/rightSection.css'
+import './styles/leftSection.css'
+import './styles/modalMedia.css'
+import './styles/media.css'
+
 import './app.js'
