+++ simple-chat/src/main.css
@@ -1,0 +1,55 @@
+*,
+*::before,
+*::after {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    outline: none;
+}
+
+:root {
+    --headerColor: #2e2e2e;
+    --white-theme: rgb(255, 255, 255);
+
+    --modalInitialMediaHeight: 0;
+    --modalInitialMediaWidth: 0;
+    --modalInitialMediaTopOffset: 0;
+    --modalInitialMediaLeftOffset: 0;
+
+    --hoverManipulationBG: rgb(230 230 230 / 40%);
+    --activeManipulationIcons: rgb(140, 140, 140)
+}
+
+li { 
+    list-style-type: none;
+}
+
+a {
+    text-decoration: none;
+}
+
+html {
+    font-size: 62.5%;
+    font-weight: 400;
+    scroll-behavior: smooth;
+}
+
+
+body {
+    font-family: Helvetica, Arial, sans-serif;
+    /* font-family: "Ubuntu", sans-serif; */
+    font-weight: 400;
+    background: #030303;
+    /* max-width: 100vw; */
+}
+
+.container {
+    position: relative;
+    width: 1280px;
+    max-width: 100%;
+    margin: 0 auto;
+    height: 100vh;
+    background-color: rgb(245, 245, 241);
+    display: flex;
+}
+
