+++ simple-chat/src/styles/contactModal.css
@@ -1,0 +1,261 @@
+
+.addContactModal.removed {
+    opacity: 0;
+}
+
+.addContactModal {
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    z-index: 1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: rgba(24, 24, 24, .3);
+    opacity: 1;
+    transition: opacity 280ms ease;
+    animation: enableDarkContactModal 200ms ease 1 forwards;
+}
+
+.addContactBlock {
+    background-color: rgb(244 244 244);
+    width: 28%;
+    height: auto;
+    max-height: 100vh;
+    /* max-height: fit-content; */
+    /* aspect-ratio: 1 / .8; */
+    border-radius: 6%;
+    -webkit-box-shadow: 0px 0px 20px 0px rgba(34, 60, 80, 0.2);
+    -moz-box-shadow: 0px 0px 20px 0px rgba(34, 60, 80, 0.2);
+    box-shadow: 0px 0px 20px 0px rgba(34, 60, 80, 0.2);
+    animation: showAddContactModal 300ms ease 1 forwards;
+    transform: translateY(60rem);
+    transition: all 800ms ease;
+}
+
+.addContactInfo {
+    width: 100%;
+    height: auto;
+    max-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    align-items: center;
+    justify-content: space-around;
+    padding: 1.8rem;
+    transition: all 800ms ease;
+}
+
+.notifyHeader {
+    /* flex: 20%; */
+    min-height: 2rem;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+}
+
+.notifyHeader p {
+    font-size: 1.4rem;
+    color: #2e2e2e;
+}
+
+.inputSection {
+    /* flex: 60%; */
+    width: 100%;
+    height: auto;
+    display: flex;
+    flex-direction: column;
+    gap: .6rem;
+    /* background-color: aqua; */
+    transition: all 800ms ease;
+}
+
+.inputSection input {
+    min-height: 4rem;
+    border: none;
+    background-color: white;
+    padding: .4rem 1rem;
+    border-radius: 1rem;
+    width: 100%;
+    position: relative;
+    z-index: 1;
+}
+
+.showSearchUser {
+    transition: background-color 200ms ease;
+    border-radius: 1rem;
+    position: relative;
+    width: 100%;
+    /* height: auto; */
+    z-index: 0;
+    height: 0;
+    opacity: 0;
+    /* transform: translateY(-3rem); */
+    /* margin-top: -5.4rem; */
+    transition: all 200ms ease;
+    /* animation: name duration timing-function delay iteration-count direction fill-mode; */
+}
+
+
+.showSearchUser.show {
+    /* transform: translateY(0); */
+    /* animation: showSearchable 280ms ease 1 forwards; */
+    height: 100%;
+    opacity: 1;
+    /* margin-top: 0; */
+}
+
+.showSearchUser:hover {
+    background-color: var(--white-theme);
+    -webkit-box-shadow: 0px 8px 30px 0px rgba(34, 60, 80, 0.2);
+    -moz-box-shadow: 0px 8px 30px 0px rgba(34, 60, 80, 0.2);
+    box-shadow: 0px 8px 30px 0px rgba(34, 60, 80, 0.2);
+}
+
+.showSearchUser.removed {
+    /* transform: translateY(-5rem); */
+    height: 0;
+    opacity: 0;
+    /* height: 0; */
+    /* animation: reverseShowSearchable 280ms ease 1 forwards; */
+    /* height: 0;
+    opacity: 0; */
+    /* transform: translateY(-3rem); */
+    /* margin-top: -5.4rem; */
+}
+
+@keyframes showSearchable {
+    0%{
+        height: 0;
+        opacity: 0;
+        display: none;
+        /* transform: translateY(-3rem); */
+        margin-top: -5.4rem;
+    }
+    98%{
+        display: block;
+    }
+    100%{
+        height: 100%;
+        opacity: 1;
+        /* transform: translateY(0); */
+        margin-top: 0;
+    }
+}
+
+@keyframes reverseShowSearchable {
+    0%{
+        opacity: 1;
+        height: 100%;
+        display: block;
+        /* transform: translateY(0); */
+    }
+    99%{
+        height: 0;
+        opacity: 0;
+        /* transform: translateY(-3rem); */
+        margin-top: -5.4rem;
+    }
+    100%{
+        display: none;
+    }
+
+}
+
+.userSection {
+    width: 100%;
+    /* height: 100%; */
+    max-height: 100vh;
+    padding: .4rem;
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+    transition: all 800ms ease;
+}
+
+.userProfileImg {
+    width: 4rem;
+    height: 4rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 50%;
+    overflow: hidden;
+}
+
+.userProfileImg img {
+    width: 100%;
+    height: 100%;
+    object-fit: contain;
+}
+
+.userDataSection {
+    display: flex;
+    align-items: center;
+}
+
+.userDataSection .userName {
+    font-size: 1.6rem;
+    color: #2e2e2e;
+}
+
+
+
+.addUserButtonSection {
+    min-height: 3.8rem;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    margin-top: .6rem;
+}
+
+.addUserButtonSection button {
+    width: 100%;
+    border: none;
+    border-radius: 1rem;
+    background-color: rgb(245, 56, 56);
+    color: white;
+    transition: background-color 280ms ease;
+}
+
+.addUserButtonSection button:hover {
+    background-color: rgb(255, 105, 105);
+}
+
+.addUserButtonSection button:active {
+    background-color: rgb(255, 176, 176);
+}
+
+.addUserButtonSection button.loading {
+    /* background-color: rgb(248, 120, 120); */
+    background-color: rgb(255, 105, 105);
+}
+
+@keyframes showAddContactModal {
+    0% {
+        transform: translateY(60rem);
+    }
+    100% {
+        transform: translateY(0);
+    }
+}
+
+@keyframes enableDarkContactModal {
+    0% {
+        background-color: rgba(0, 0, 0, 0);
+    }
+    100% {
+        background-color: rgba(0, 0, 0, .1);
+    }
+}
+
+
+@media screen and (max-width: 425px) {
+    .addContactModal {
+        z-index: 5;
+    }
+
+    .addContactBlock {
+        width: 80%;
+    }
+}
