+++ simple-chat/src/styles/modalMedia.css
@@ -1,0 +1,251 @@
+/* модалка */
+
+
+
+
+.contentBlock {
+    display: flex;
+    justify-content: space-between;
+    /* background-color: bisque; */
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    padding: 1.2rem 1.6rem;
+    z-index: 3;
+    opacity: 1;
+    transition: opacity .8s;
+}
+
+.contentBlock.hide {
+    opacity: 0;
+}
+
+.leftSideBlock {
+    min-width: 16%;
+    max-width: 16%;
+    display: flex;
+    align-items: center;
+    /* background-color: #7a7a7a; */
+    align-self: flex-end;
+    padding: .4rem 0;
+}
+
+.contentDataBlock {
+    display: flex;
+    gap: 1rem;
+    font-size: 1.4rem;
+    color: rgb(170, 170, 170);
+}
+
+.textDataBlock {
+    max-width: 60%;
+    background-color: rgba(40, 40, 40, 0.7);
+    padding: .8rem 1.2rem;
+    border-radius: .8rem;
+    font-size: 1.4rem;
+    color: var(--white-theme);
+}
+
+.textDataBlock p {
+    text-align: center;
+    line-height: 1.8rem;
+}
+
+.rightSideBlock {
+    min-width: 16%;
+    max-width: 16%;
+    display: flex;
+    justify-content: flex-end;
+    align-self: flex-end;
+    padding: .4rem 0;
+}
+
+.manipulationStatus {
+    display: flex;
+    height: 2rem;
+    width: 2rem;
+    align-items: center;
+    justify-content: center;
+    font-size: 1.4rem;
+    color: rgb(170, 170, 170);
+}
+
+.manipulationStatus span {
+    font-size: 2.4rem;
+}
+
+
+
+/* модалка на смартфоны */
+
+
+.modalMedia {
+    display: none;
+}
+
+.modalMedia.removed {
+   opacity: 0;
+}
+
+.modalMedia.visible {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100vh;
+    background-color: #000000;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    z-index: 10;
+    transition: opacity .2s;
+}
+
+/* .modalImg .modalImgWrapper {
+    position: fixed;
+} */
+
+.modalMedia .modalMedia_Header.visible {
+    opacity: 1;
+    top: 0;
+}
+
+.modalMedia .modalMedia_Header .modalHeader_left {
+    display: flex;
+    align-items: center;
+    gap: 2rem;
+}
+
+.modalMedia .modalMedia_Header .modalHeader_left .modalMedia_Header_back {
+    display: flex;
+    width: 2.8rem;
+    height: 2.8rem;
+    border-radius: 50%;
+    justify-content: center;
+    align-items: center;
+    transition: all 140ms ease;
+}
+
+.modalMedia .modalMedia_Header .modalHeader_left .modalMedia_Header_back:active {
+    background-color: rgb(110 110 110 / 80%);
+}
+
+.modalMedia .modalMedia_Header .modalHeader_left .modalMedia_Header_back span {
+    font-size: 2rem;
+}
+
+.modalMedia .modalMedia_Header .modalHeader_left .modalMedia_Header_chatData {
+    display: flex;
+    flex-direction: column;
+    gap: .3rem;
+}
+
+.modalMedia .modalMedia_Header .modalHeader_left .modalMedia_Header_chatData .modalMedia_Header_chatData_name {
+    font-weight: 500;
+    font-size: 1.6rem;
+}
+
+.modalMedia .modalMedia_Header .modalHeader_left .modalMedia_Header_chatData .modalMedia_Header_chatData_date {
+    font-weight: 400;
+    font-size: 1.3rem;
+    color: rgb(230, 230, 230);
+}
+.modalMedia .modalMedia_Header .modalMedia_Header_right {
+    display: flex;
+    height: 2.8rem;
+    align-items: center;
+    gap: 1rem;
+}
+
+.modalMedia .modalMedia_Header .modalMedia_Header_right span {
+    font-size: 1.8rem;
+    border-radius: 50%;
+}
+
+.modalMedia .modalMedia_Header .modalMedia_Header_right span:active {
+    background-color: rgb(110 110 110 / 80%);
+}
+
+.modalMedia .modalMedia_Header {
+    position: fixed;
+    top: -4rem;
+    display: flex;
+    color: #fff;
+    width: 100%;
+    padding: .6rem 1rem;
+    font-size: 1.4rem;
+    justify-content: space-between;
+    z-index: 4;
+    align-items: center;
+    opacity: 0;
+    transition: opacity 200ms ease, top 200ms ease;
+}
+
+.modalMedia .modalMedia_content {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    z-index: 3;
+    position: absolute;
+}
+
+.modalMedia .modalMedia_content .img_wrapper {
+    width: 100%;
+    max-height: 100%;
+}
+
+
+.modalMedia .modalMedia_content .img_wrapper img.desktop {
+    display: block;
+    width: 100%;
+    height: 100%;
+    object-fit: contain;
+}
+
+
+.modalMedia .modalMedia_content .img_wrapper img.mobile {
+    display: block;
+    position: absolute;
+    height: var(--modalInitialMediaHeight);
+    width: var(--modalInitialMediaWidth);
+    top: var(--modalInitialMediaTopOffset);
+    left: var(--modalInitialMediaLeftOffset);
+    object-fit: contain;
+    transition: all 180ms ease-out;
+}
+
+.modalMedia .modalMedia_content .img_wrapper img.initial {
+    /* position: static; */
+    width: 100%;
+    height: 100%;
+}
+
+.modalMedia .modalMedia_text {
+    position: fixed;
+    bottom: -4rem;
+    color: #fff;
+    width: 100%;
+    padding: 1rem 1.4rem 1.6rem;
+    background-color: #00000080;
+    z-index: 4;
+    opacity: 0;
+    transition: opacity 200ms ease, bottom 200ms ease;
+}
+
+.modalMedia .modalMedia_text.visible {
+    /* position: fixed;
+    bottom: 0;
+    color: #fff;
+    width: 100%;
+    padding: 1rem 1.4rem 1.6rem;
+    background-color: #00000080;
+    z-index: 4; */
+    opacity: 1;
+    bottom: 0;
+}
+
+.modalMedia .modalMedia_text p {
+    font-size: 1.5rem;
+}
